Reorganize MTL
Now we have separate source sets for each directory and a clear dependency
structure. Also, add some utilities that are used by the dart_content_handler.

Fix a bug in MessageLoop where it won't drain if it thought a drain was already
scheduled. Now that we wait on handles between reloading the priority queue, we
need to signal the message loop whenever the incoming task queue transition
from empty to non-empty.

Change-Id: I1cd3b32185458dc5353dd53712e2dd0bac9af603